Refactor Apple Store clone with improved structure and UI components
