--- apple-store-clone/app/(store)/layout.tsx
@@ -1,33 +1,0 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "@/app/globals.css";
-import { NavBar } from "@/components/NavBar";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: " Store (clone)",
-  description: "Clone Apple Store avec Next.js",
-};
-
-export default function StoreLayout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  return (
-    <html lang="fr">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <NavBar />
-        <main className="mx-auto max-w-7xl px-4 py-8">{children}</main>
-      </body>
-    </html>
-  );
-}
-
